更新数据库文档：t_ws_bill_detail商品数量改为decimal(19, 8)

+++ doc/99 SQL/01CreateTables.sql
@@ -440,7 +440,7 @@ CREATE TABLE IF NOT EXISTS `t_ws_bill_detail` (
   `id` varchar(255) NOT NULL,
   `date_created` datetime DEFAULT NULL,
   `goods_id` varchar(255) NOT NULL,
-  `goods_count` int(11) NOT NULL,
+  `goods_count` decimal(19,8) NOT NULL,
   `goods_money` decimal(19,2) NOT NULL,
   `goods_price` decimal(19,2) NOT NULL,
   `inventory_money` decimal(19,2) DEFAULT NULL,
